+++ src/main/java/ru/school/retailanalitycs_web_java/exceptions/notFoundExceptions/SkuNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.school.retailanalitycs_web_java.exceptions.notFoundExceptions;
 
 public class SkuNotFoundException extends EntityNotFoundException {
-    public SkuNotFoundException(Integer id) {
+    public SkuNotFoundException(Long id) {
         super(String.format("Sku with id {%s} not found", id));
     }
 }
