+++ src/main/java/ru/school/retailanalitycs_web_java/exceptions/notFoundExceptions/TransactionNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.school.retailanalitycs_web_java.exceptions.notFoundExceptions;
 
 public class TransactionNotFoundException extends EntityNotFoundException {
-    public TransactionNotFoundException(Integer id) {
+    public TransactionNotFoundException(Long id) {
         super(String.format("Transaction with id {%s} not found", id));
     }
 }
