+++ src/main/java/ru/school/retailanalitycs_web_java/repositories/entityRepositories/SkuRepository.java
@@ -3,5 +3,5 @@ package ru.school.retailanalitycs_web_java.repositories.entityRepositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.school.retailanalitycs_web_java.entities.tables.Sku;
 
-public interface SkuRepository extends JpaRepository<Sku, Integer> {
+public interface SkuRepository extends JpaRepository<Sku, Long> {
 }
