+++ src/main/java/ru/school/retailanalitycs_web_java/repositories/viewRepositories/CustomerViewRepository.java
@@ -2,5 +2,5 @@ package ru.school.retailanalitycs_web_java.repositories.viewRepositories;
 
 import ru.school.retailanalitycs_web_java.entities.views.CustomerView;
 
-public interface CustomerViewRepository extends ReadOnlyRepository<CustomerView, Integer> {
+public interface CustomerViewRepository extends ReadOnlyRepository<CustomerView, Long> {
 }
